+++ fpga-runtime/src/frt/devices/xilinx_opencl_device.cpp
@@ -317,7 +317,7 @@ std::unique_ptr<Device> XilinxOpenclDevice::New(
   return std::make_unique<XilinxOpenclDevice>(binaries);
 }
 
-void XilinxOpenclDevice::SetStreamArg(int index, Tag tag, StreamArg& arg) {
+void XilinxOpenclDevice::SetStreamArg(size_t index, Tag tag, StreamArg& arg) {
   LOG(FATAL) << "Xilinx OpenCL streaming is disabled";
 }
 
@@ -343,7 +343,7 @@ void XilinxOpenclDevice::ReadFromDevice() {
   }
 }
 
-cl::Buffer XilinxOpenclDevice::CreateBuffer(int index, cl_mem_flags flags,
+cl::Buffer XilinxOpenclDevice::CreateBuffer(size_t index, cl_mem_flags flags,
                                             void* host_ptr, size_t size) {
   flags |= CL_MEM_USE_HOST_PTR;
   return OpenclDevice::CreateBuffer(index, flags, host_ptr, size);
